+++ internal/contextkeys/contextkeys.go
@@ -5,10 +5,7 @@ import (
 	"context"
 	"time"
 
-	"github.com/gin-gonic/gin"
-	"github.com/gofiber/fiber/v2"
 	"github.com/golang-jwt/jwt/v5"
-	"github.com/hewen/mastiff-go/server/httpx/unicontext"
 )
 
 const (
@@ -108,45 +105,4 @@ func SetRedisBeginTime(ctx context.Context, t time.Time) context.Context {
 func GetRedisBeginTime(ctx context.Context) (time.Time, bool) {
 	t, ok := ctx.Value(RedisBeginTimeKey).(time.Time)
 	return t, ok
-}
-
-// ContextFrom extracts context.Context from gin.Context, fiber.Ctx or context.Context itself.
-// If none matched, returns context.Background.
-func ContextFrom(v any) context.Context {
-	// NOTE: Order matters in type switch — match *gin.Context and *fiber.Ctx
-	// before context.Context to avoid premature capture.
-	switch c := v.(type) {
-	case unicontext.UniversalContext:
-		if val, ok := c.Get(ContextKey); ok && val != nil {
-			if ctx, ok := val.(context.Context); ok {
-				return ctx
-			}
-		}
-	case *gin.Context:
-		if req := c.Request; req != nil {
-			return req.Context()
-		}
-	case *fiber.Ctx:
-		if val := c.Locals(ContextKey); val != nil {
-			if ctx, ok := val.(context.Context); ok {
-				return ctx
-			}
-		}
-	case context.Context:
-		return c
-	}
-
-	return context.Background()
-}
-
-// InjectContext sets the updated context.Context back into the carrier (gin/fiber).
-func InjectContext(ctx context.Context, carrier any) {
-	switch c := carrier.(type) {
-	case unicontext.UniversalContext:
-		c.Set(ContextKey, ctx)
-	case *gin.Context:
-		c.Request = c.Request.WithContext(ctx)
-	case *fiber.Ctx:
-		c.Locals(ContextKey, ctx)
-	}
 }
