+++ internal/contextkeys/contextkeys_test.go
@@ -2,16 +2,11 @@ package contextkeys
 
 import (
 	"context"
-	"net/http"
 	"testing"
 	"time"
 
-	"github.com/gin-gonic/gin"
-	"github.com/gofiber/fiber/v2"
 	"github.com/golang-jwt/jwt/v5"
-	"github.com/hewen/mastiff-go/server/httpx/unicontext"
 	"github.com/stretchr/testify/assert"
-	"github.com/valyala/fasthttp"
 )
 
 func TestCtxKeyString(t *testing.T) {
@@ -87,123 +82,4 @@ func TestGenericSetAndGetValue(t *testing.T) {
 	assert.True(t, ok)
 	assert.Equal(t, "baz", got.Foo)
 	assert.Equal(t, 42, got.Bar)
-}
-
-func TestContextFrom_GinContext(t *testing.T) {
-	gin.SetMode(gin.TestMode)
-	ctx := context.WithValue(context.Background(), LoggerTraceIDKey, "12345")
-
-	// mock gin.Context
-	req, _ := http.NewRequest(http.MethodGet, "/", nil)
-	req = req.WithContext(ctx)
-	c := &gin.Context{
-		Request: req,
-	}
-
-	got := ContextFrom(c)
-	traceID, ok := GetTraceID(got)
-
-	assert.True(t, ok)
-	assert.Equal(t, "12345", traceID)
-}
-
-func TestContextFrom_FiberCtx(t *testing.T) {
-	app := fiber.New()
-	testCtx := context.WithValue(context.Background(), LoggerTraceIDKey, "abcde")
-
-	c := app.AcquireCtx(&fasthttp.RequestCtx{})
-	c.Locals(ContextKey, testCtx)
-
-	got := ContextFrom(c)
-	traceID, ok := GetTraceID(got)
-
-	assert.True(t, ok)
-	assert.Equal(t, "abcde", traceID)
-}
-
-func TestContextFrom_Unicontext(t *testing.T) {
-	testCtx := context.WithValue(context.Background(), LoggerTraceIDKey, "abcde")
-
-	app := fiber.New()
-	ctx := &unicontext.FiberContext{
-		Ctx: app.AcquireCtx(&fasthttp.RequestCtx{}),
-	}
-	ctx.Set(ContextKey, testCtx)
-
-	got := ContextFrom(ctx)
-	traceID, ok := GetTraceID(got)
-
-	assert.True(t, ok)
-	assert.Equal(t, "abcde", traceID)
-}
-
-func TestContextFrom_Context(t *testing.T) {
-	ctx := context.WithValue(context.Background(), LoggerTraceIDKey, "xyz")
-
-	got := ContextFrom(ctx)
-	traceID, ok := GetTraceID(got)
-
-	assert.True(t, ok)
-	assert.Equal(t, "xyz", traceID)
-}
-
-func TestContextFrom_UnknownType(t *testing.T) {
-	got := ContextFrom("unknown")
-	assert.NotNil(t, got)
-	assert.Equal(t, context.Background(), got)
-}
-
-func TestInjectContext_Gin(t *testing.T) {
-	gin.SetMode(gin.TestMode)
-
-	ctx := context.WithValue(context.Background(), LoggerTraceIDKey, "gin-trace")
-	req, _ := http.NewRequest(http.MethodGet, "/", nil)
-	c := &gin.Context{Request: req}
-
-	InjectContext(ctx, c)
-
-	got := c.Request.Context()
-	traceID, ok := GetTraceID(got)
-	assert.True(t, ok)
-	assert.Equal(t, "gin-trace", traceID)
-}
-
-func TestInjectContext_Fiber(t *testing.T) {
-	app := fiber.New()
-	ctx := context.WithValue(context.Background(), LoggerTraceIDKey, "fiber-trace")
-	c := app.AcquireCtx(&fasthttp.RequestCtx{})
-
-	InjectContext(ctx, c)
-
-	val := c.Locals(ContextKey)
-	assert.NotNil(t, val)
-
-	retrieved, ok := val.(context.Context)
-	assert.True(t, ok)
-
-	traceID, found := GetTraceID(retrieved)
-	assert.True(t, found)
-	assert.Equal(t, "fiber-trace", traceID)
-}
-
-func TestInjectContext_Unicontext(t *testing.T) {
-	app := fiber.New()
-	ctx := context.WithValue(context.Background(), LoggerTraceIDKey, "fiber-trace")
-
-	c := &unicontext.FiberContext{
-		Ctx: app.AcquireCtx(&fasthttp.RequestCtx{}),
-	}
-
-	InjectContext(ctx, c)
-
-	val, ok := c.Get(ContextKey)
-	assert.NotNil(t, val)
-	assert.True(t, ok)
-
-	retrieved, ok := val.(context.Context)
-	assert.True(t, ok)
-
-	traceID, found := GetTraceID(retrieved)
-	assert.True(t, found)
-	assert.Equal(t, "fiber-trace", traceID)
 }
