+++ middleware/recovery/httpx.go
@@ -5,7 +5,6 @@ import (
 	"runtime/debug"
 	"strings"
 
-	"github.com/hewen/mastiff-go/internal/contextkeys"
 	"github.com/hewen/mastiff-go/logger"
 	"github.com/hewen/mastiff-go/server/httpx/unicontext"
 )
@@ -15,7 +14,7 @@ func HttpxMiddleware() func(c unicontext.UniversalContext) error {
 	return func(c unicontext.UniversalContext) error {
 		defer func() {
 			if r := recover(); r != nil {
-				l := logger.NewLoggerWithContext(contextkeys.ContextFrom(c))
+				l := logger.NewLoggerWithContext(unicontext.ContextFrom(c))
 				l.Errorf("panic: %v $%s", r, strings.ReplaceAll(string(debug.Stack()), "\n", "$"))
 			}
 		}()
