+++ server/httpx/handler/fiber_test.go
@@ -10,7 +10,6 @@ import (
 
 	"github.com/hewen/mastiff-go/config/middlewareconf"
 	"github.com/hewen/mastiff-go/config/serverconf"
-	"github.com/hewen/mastiff-go/internal/contextkeys"
 	"github.com/hewen/mastiff-go/logger"
 	"github.com/hewen/mastiff-go/middleware/logging"
 	"github.com/hewen/mastiff-go/pkg/util"
@@ -41,7 +40,7 @@ func TestFiberHandler(t *testing.T) {
 	s.Use(logging.HttpxMiddleware())
 
 	s.Get("/test", func(c unicontext.UniversalContext) error {
-		ctx := contextkeys.ContextFrom(c)
+		ctx := unicontext.ContextFrom(c)
 		l := logger.NewLoggerWithContext(ctx)
 		l.Infof("test")
 		return c.JSON(http.StatusOK, map[string]string{
